+++ DSCResources/MSFT_CertReq/MSFT_CertReq.psm1
@@ -76,6 +76,12 @@ $localizedData = Get-LocalizedData `
 
     .PARAMETER FriendlyName
     Specifies a friendly name for the certificate.
+
+    .PARAMETER KeyType
+    Specifies if the key type should be RSA or ECDH, defaults to RSA.
+
+    .PARAMETER RequestType
+    Specified if the request type should be CMC or PKCS10, deafults to CMC.
 #>
 function Get-TargetResource
 {
@@ -97,7 +103,7 @@ function Get-TargetResource
         $CARootName,
 
         [Parameter()]
-        [ValidateSet("1024", "2048", "4096", "8192")]
+        [ValidateSet('192', '224', '256', '384', '521', '1024', '2048', '4096', '8192')]
         [System.String]
         $KeyLength = '2048',
 
@@ -160,8 +166,20 @@ function Get-TargetResource
         [Parameter()]
         [ValidateNotNullOrEmpty()]
         [System.String]
-        $FriendlyName
+        $FriendlyName,
+
+        [Parameter()]
+        [ValidateSet('RSA', 'ECDH')]
+        [System.String]
+        $KeyType = 'RSA',
+
+        [Parameter()]
+        [ValidateSet('CMC', 'PKCS10')]
+        [System.String]
+        $RequestType = 'CMC'
     )
+
+    Assert-ResourceProperty -KeyLength $KeyLength -KeyType $KeyType
 
     # The certificate authority, accessible on the local area network
     if ([string]::IsNullOrWhiteSpace($CAServerFQDN) -or [string]::IsNullOrWhiteSpace($CARootName))
@@ -272,6 +290,12 @@ function Get-TargetResource
 
     .PARAMETER FriendlyName
     Specifies a friendly name for the certificate.
+
+    .PARAMETER KeyType
+    Specifies if the key type should be RSA or ECDH, defaults to RSA.
+
+    .PARAMETER RequestType
+    Specified if the request type should be CMC or PKCS10, deafults to CMC.
 #>
 function Set-TargetResource
 {
@@ -292,7 +316,7 @@ function Set-TargetResource
         $CARootName,
 
         [Parameter()]
-        [ValidateSet("1024", "2048", "4096", "8192")]
+        [ValidateSet('192', '224', '256', '384', '521', '1024', '2048', '4096', '8192')]
         [System.String]
         $KeyLength = '2048',
 
@@ -355,8 +379,20 @@ function Set-TargetResource
         [Parameter()]
         [ValidateNotNullOrEmpty()]
         [System.String]
-        $FriendlyName
+        $FriendlyName,
+
+        [Parameter()]
+        [ValidateSet('RSA', 'ECDH')]
+        [System.String]
+        $KeyType = 'RSA',
+
+        [Parameter()]
+        [ValidateSet('CMC', 'PKCS10')]
+        [System.String]
+        $RequestType = 'CMC'
     )
+
+    Assert-ResourceProperty -KeyLength $KeyLength -KeyType $KeyType
 
     # The certificate authority, accessible on the local area network
     if ([string]::IsNullOrWhiteSpace($CAServerFQDN) -or [string]::IsNullOrWhiteSpace($CARootName))
@@ -409,7 +445,6 @@ function Set-TargetResource
     $userProtected = 'FALSE'
     $useExistingKeySet = 'FALSE'
     $providerType = '12'
-    $requestType = 'CMC'
 
     # A unique identifier for temporary files that will be used when interacting with the command line utility
     $guid = [system.guid]::NewGuid().guid
@@ -432,7 +467,7 @@ UserProtected = $userProtected
 UseExistingKeySet = $useExistingKeySet
 ProviderName = $ProviderName
 ProviderType = $providerType
-RequestType = $requestType
+RequestType = $RequestType
 KeyUsage = $KeyUsage
 "@
     if ($FriendlyName)
@@ -477,6 +512,7 @@ CertificateTemplate = "$CertificateTemplate"
 2.5.29.17 = "{text}$SubjectAltName"
 "@
     }
+
     Set-Content -Path $infPath -Value $requestDetails
 
     <#
@@ -547,8 +583,8 @@ CertificateTemplate = "$CertificateTemplate"
                     '-p', [System.Runtime.InteropServices.Marshal]::PtrToStringAuto($credPW),
                     '-PolicyServer', $CepURL,
                     '-config', $CesURL,
-                    $ReqPath,
-                    $CerPath
+                    $reqPath,
+                    $cerPath
                 )
             }
             else
@@ -607,7 +643,7 @@ CertificateTemplate = "$CertificateTemplate"
         }
         else
         {
-            $submitRequest = & certreq.exe @('-submit', '-q', '-config', $CA, $ReqPath, $CerPath)
+            $submitRequest = & certreq.exe @('-submit', '-q', '-config', $CA, $reqPath, $cerPath)
         } # if
 
         Write-Verbose -Message ( @(
@@ -703,6 +739,12 @@ CertificateTemplate = "$CertificateTemplate"
 
     .PARAMETER FriendlyName
     Specifies a friendly name for the certificate.
+
+    .PARAMETER KeyType
+    Specifies if the key type should be RSA or ECDH, defaults to RSA.
+
+    .PARAMETER RequestType
+    Specified if the request type should be CMC or PKCS10, deafults to CMC.
 #>
 function Test-TargetResource
 {
@@ -724,7 +766,7 @@ function Test-TargetResource
         $CARootName,
 
         [Parameter()]
-        [ValidateSet("1024", "2048", "4096", "8192")]
+        [ValidateSet('192', '224', '256', '384', '521', '1024', '2048', '4096', '8192')]
         [System.String]
         $KeyLength = '2048',
 
@@ -787,8 +829,20 @@ function Test-TargetResource
         [Parameter()]
         [ValidateNotNullOrEmpty()]
         [System.String]
-        $FriendlyName
+        $FriendlyName,
+
+        [Parameter()]
+        [ValidateSet('RSA', 'ECDH')]
+        [System.String]
+        $KeyType = 'RSA',
+
+        [Parameter()]
+        [ValidateSet('CMC', 'PKCS10')]
+        [System.String]
+        $RequestType = 'CMC'
     )
+
+    Assert-ResourceProperty -KeyLength $KeyLength -KeyType $KeyType
 
     # The certificate authority, accessible on the local area network
     if ([string]::IsNullOrWhiteSpace($CAServerFQDN) -or [string]::IsNullOrWhiteSpace($CARootName))
@@ -841,7 +895,7 @@ function Test-TargetResource
 
         if ($AutoRenew)
         {
-            if ($Cert.NotAfter -le (Get-Date).AddDays(-30))
+            if ($Cert.NotAfter -le (Get-Date).AddDays(30))
             {
                 # The certificate was found but it is expiring within 30 days or has expired
                 Write-Verbose -Message ( @(
@@ -940,6 +994,39 @@ function Test-TargetResource
         ) -join '' )
     return $false
 } # end function Test-TargetResource
+
+<#
+    .SYNOPSIS
+    This function will check and ensure the right key length was choosen for the key type that was intended to be used
+
+    .PARAMETER KeyLength
+    The bit length of the encryption key to be used.
+
+    .PARAMETER KeyType
+    Specifies if the key type should be RSA or ECDH, defaults to RSA.
+#>
+function Assert-ResourceProperty
+{
+    [CmdletBinding()]
+    param
+    (
+        [Parameter()]
+        [ValidateSet('192', '224', '256', '384', '521', '1024', '2048', '4096', '8192')]
+        [System.String]
+        $KeyLength = '2048',
+
+        [Parameter()]
+        [ValidateSet('RSA', 'ECDH')]
+        [System.String]
+        $KeyType = 'RSA'
+    )
+
+    if ((($KeyType -eq 'RSA') -and ($KeyLength -notin '1024', '2048', '4096', '8192')) -or `
+    (($KeyType -eq 'ECDH') -and ($KeyLength -notin '192', '224', '256', '384', '521')))
+    {
+        New-InvalidArgumentException -Message (($LocalizedData.InvalidKeySize) -f $KeyLength,$KeyType) -ArgumentName 'KeyLength'
+    }
+}# end function Assert-ResourceProperty
 
 <#
     .SYNOPSIS