+++ CHANGELOG.md
@@ -24,7 +24,12 @@
 - CertificateExport:
   - Fixed bug causing PFX export with matchsource enabled to fail - fixes
     [Issue 117](https://github.com/PowerShell/CertificateDsc/issues/117)
+- Added DSCResourcesToExport to the CertificateDSC.psd1
 - CertReq:
+  - Added key lengths for ECDH key type.
+  - Added Key type to check for correct key lengths. - fixes
+    [Issue 113](https://github.com/PowerShell/CertificateDsc/issues/113)
+  - Added request type parameter to support PKCS10.
   - Simplified unit test comparison certificate request strings to make
     tests easier to read.
   - Improved unit test layout and updated to meet standards.